Working refactor
The only difference with the original scripts is that
the subsetting to a specific time-window isn't applied consistently
in the original scripts, which can result in different background levels
etc... and thus impact on the detected plumes, which is the case for
SO2.

I should confirm with the code authors the purpose of this subsetting
and whether it can be applied upfront.