Removed extended plumes estimation
Previously plumes were found by values that are 3 SDs above the
background, then the start time was pushed back to when the
concentration went 2 SDs above the background so as to not lose too much
of the plume. However, this peeking back only added a few hundred
milliseconds to the plume, which was subsequently dwarfed by extending
the plumes with a 5s buffer.

Therefore, to save calculations that ultimately don't impact on the
plume estimation, this extended plume estimation has been removed.