Optimised reduction of overlapping plumes
Previously this was done by finding all unique plume times, then
grouping these again. Now, a recursive function is used to identify the
set of non-overlapping plume (start,end) times.

Furthermore, the buffer time is now applied just to one plume and only
when checking to see if the plume is overlapping. Previously 5s was just
added to either side of a plume, meaning that both plumes were
artificially widened, but also that plumes were considered overlapping
if they overlapped by 10s, not the 5s that the parameter implies.
Now, the buffer is only used when checking for overlappingness, and
doesn't otherwise change the plume window times