Reverted back to providing background to integrate_aup_trapz
This means that there is no scenario in which the background is removed.
The previous idea of using the output from the Wavelet HPF would be
nice, but it loses some of the full range from the peaks, meaning the
areas are smaller than they should be.
Instead the wavelets will just be used to find the peak *times*, rather
than the full values as well