feat: make background removal explicit when calculating areas (#9)
* Reduce down to 1 row per plume if not using uncertainty

* `integrate_aup_trapz` automatically subtracts background

This removes the onus on the user to do it themselves, and is flexible
enough to handle either a manually provided background or will linearly
interpolate one if not provided.

* Bumped version number

* Added automatic background removal from plume area estimation

* Updated to np.trapezoid (np.trapz is deprecated)

* Updated README

* Fixed issue with plots overwriting

* Refactored `detect_plumes_wavelet` to return concetration & reconstructed signal

Reconstructed signal is the signal without the background, as obtained
from the HPF

* Soft deprecated old rolling window detection method

* Added mention of `plume_buffer`

* Reverted back to providing background to integrate_aup_trapz

This means that there is no scenario in which the background is removed.
The previous idea of using the output from the Wavelet HPF would be
nice, but it loses some of the full range from the peaks, meaning the
areas are smaller than they should be.
Instead the wavelets will just be used to find the peak *times*, rather
than the full values as well

* Specify the number of levels, rather than a list of levels

The only benefit to passing in the last of coefficients is that you can
use non-contiguous ones, like say 12, 14, 17.
But in reality, you'd always want to use a continuous set.

* Bumped version number